Fix "Open with default file handler on Linux"
As `open` exists on Linux machines, the first if statement evaluated
to true, so `open` was used on all systems. Switching if branches, and
testing on `pbpaste` first, and `open` last does the trick for now.

There might still be problem if a Linuc machine doesn't have `xdg-open`
but does have `open`. But I deal with this at some other time.