Find process via search and kill with menu
From

http://brettterpstra.com/2009/11/14/fk-a-useful-bash-function/

It lets you do a quick search of your running processes for a
case-insensitive partial match of the first parameter passed to it.
It's useful for quickly finding a process without worrying about its
capitalization or full spelling, and without having to sift through (or
manually grep) a long ps ax list.

I just added some input validation, moved the Cancel option to the top
of the menu, so that it always is the same, and return the function
directly if no rsult has been found.