Create empty jline settings to NOT screw with sbt
The newer versions of sbt/scala repl use more curent version of Jline
that reads and recognizes keybindings in ~/.inputrc like

  "\e[A":history-search-backward
  "\e[B":history-search-forward

That will move forward and backward in the history. This screws with the
repl though.

The solution is to create a .jline-inputrc,and refer to it via .jlinerc
which will take precedence.

Thanks to Paul Phillips who posted the solution to the mailing list.

https://groups.google.com/d/msg/simple-build-tool/ShikT6VAd_g/uvH1DCPdS3IJ